[UKET-47] 리패키징 (#3)
