--- src/main/kotlin/application/port/TestPort.kt
@@ -1,3 +1,0 @@
-package uket.application.port
-
-interface TestPort
