--- src/main/kotlin/application/usecase/TestUseCase.kt
@@ -1,3 +1,0 @@
-package uket.application.usecase
-
-interface TestUseCase
