--- src/main/kotlin/domain/Test.kt
@@ -1,3 +1,0 @@
-package uket.domain
-
-class Test
