--- src/main/kotlin/gateway/exception/BaseException.kt
@@ -1,3 +1,0 @@
-package uket.gateway.exception
-
-data class BaseException(val errorCode: ErrorCode) : RuntimeException(errorCode.message)
