--- src/main/kotlin/gateway/interceptor/LoginInterceptor.kt
@@ -1,56 +1,0 @@
-package uket.gateway.interceptor
-
-import com.fasterxml.jackson.databind.ObjectMapper
-import jakarta.servlet.http.HttpServletRequest
-import jakarta.servlet.http.HttpServletResponse
-import org.springframework.http.HttpMethod
-import org.springframework.http.MediaType
-import org.springframework.stereotype.Component
-import org.springframework.web.servlet.HandlerInterceptor
-import uket.gateway.exception.BaseException
-import uket.gateway.exception.ErrorCode
-import uket.gateway.exception.ErrorResponse
-import uket.gateway.interceptor.AuthorizationExtractor.extractAccessToken
-import uket.gateway.jwt.filter.TokenValidator
-import java.io.IOException
-import java.nio.charset.StandardCharsets
-
-@Component
-class LoginInterceptor(
-    private val tokenValidator: TokenValidator,
-    private val objectMapper: ObjectMapper,
-) : HandlerInterceptor {
-    override fun preHandle(request: HttpServletRequest, response: HttpServletResponse, handler: Any): Boolean {
-        if (isPreflight(request) || isSwaggerRequest(request)) {
-            return true
-        }
-
-        val token = extractAccessToken(request)
-        try {
-            tokenValidator.validateRegistered(token)
-        } catch (exception: BaseException) {
-            val errorCode: ErrorCode = exception.errorCode
-            writeErrorResponse(response, errorCode)
-            return false
-        }
-        return true
-    }
-
-    @Throws(IOException::class)
-    private fun writeErrorResponse(response: HttpServletResponse, errorCode: ErrorCode) {
-        response.characterEncoding = StandardCharsets.UTF_8.name()
-        response.contentType = MediaType.APPLICATION_JSON_VALUE
-        response.status = HttpServletResponse.SC_UNAUTHORIZED
-        response.writer
-            .write(objectMapper.writeValueAsString(ErrorResponse.of(errorCode)))
-    }
-
-    private fun isSwaggerRequest(request: HttpServletRequest): Boolean {
-        val uri = request.requestURI
-        return uri.contains("swagger") || uri.contains("api-docs") || uri.contains("webjars")
-    }
-
-    private fun isPreflight(request: HttpServletRequest): Boolean {
-        return request.method == HttpMethod.OPTIONS.toString()
-    }
-}
