--- src/main/kotlin/gateway/jwt/filter/AuthenticationUserInfo.kt
@@ -1,8 +1,0 @@
-package uket.gateway.jwt.filter
-
-data class AuthenticationUserInfo(
-    val userId: Long,
-    val name: String,
-    val role: String,
-    val isRegistered: Boolean = false,
-)
