--- src/main/kotlin/infra/output/persistence/adapter/TestAdapter.kt
@@ -1,7 +1,0 @@
-package uket.infra.output.persistence.adapter
-
-import org.springframework.stereotype.Component
-import uket.application.port.TestPort
-
-@Component
-class TestAdapter : TestPort
