--- src/main/kotlin/infra/output/persistence/entity/TestEntity.kt
@@ -1,3 +1,0 @@
-package uket.infra.output.persistence.entity
-
-class TestEntity
