+++ src/main/kotlin/uket/auth/exception/BaseException.kt
@@ -1,0 +1,3 @@
+package uket.auth.exception
+
+data class BaseException(val errorCode: ErrorCode) : RuntimeException(errorCode.message)
