+++ src/main/kotlin/uket/auth/exception/ErrorResponse.kt
@@ -1,0 +1,16 @@
+package uket.auth.exception
+
+data class ErrorResponse(
+    val code: String,
+    val message: String,
+) {
+    companion object {
+        fun of(code: String, message: String): ErrorResponse {
+            return ErrorResponse(code, message)
+        }
+
+        fun of(errorCode: ErrorCode, exception: Throwable? = null): ErrorResponse {
+            return ErrorResponse(errorCode.code, exception?.message ?: errorCode.message)
+        }
+    }
+}
