+++ src/main/kotlin/uket/auth/filter/AuthenticationUserInfo.kt
@@ -1,0 +1,8 @@
+package uket.auth.filter
+
+data class AuthenticationUserInfo(
+    val userId: Long,
+    val name: String,
+    val role: String,
+    val isRegistered: Boolean = false,
+)
