+++ src/main/kotlin/uket/auth/interceptor/AuthorizationExtractor.kt
@@ -1,0 +1,17 @@
+package uket.auth.interceptor
+
+import jakarta.servlet.http.HttpServletRequest
+import uket.auth.exception.BaseException
+import uket.auth.exception.ErrorCode
+import uket.uket.auth.jwt.JwtValues.JWT_AUTHORIZATION_HEADER
+import uket.uket.auth.jwt.JwtValues.JWT_AUTHORIZATION_VALUE_PREFIX
+
+object AuthorizationExtractor {
+    @JvmStatic
+    fun extractAccessToken(request: HttpServletRequest): String {
+        val accessToken = request.getHeader(JWT_AUTHORIZATION_HEADER)
+            ?: throw BaseException(ErrorCode.AUTHENTICATION_FAILED)
+
+        return accessToken.replace(JWT_AUTHORIZATION_VALUE_PREFIX, "")
+    }
+}
