+++ src/main/kotlin/uket/domain/organization/Organization.kt
@@ -1,0 +1,3 @@
+package uket.uket.domain.organization
+
+class Organization
